@mccord 2020-06-23
> Nice! Also were you able to retest the perf results after fixing the batch/key param issue?
> I think that before, in __prepOneKey, key was always undefined, which means we always returned the same valueIndex in that function? Not sure how that might affect things

Ah yes. Sorry, I was switching between coding in this repo and coding in a temp (pre-built) file which I would c sync to cloud-dev node modules, so the code in this PR wasn't actually reflecting the code that was running.