@mccord 2020-06-23
See https://github.com/ThreadsHQ/Web/pull/7471/files instead